requirements.txt
pyautogui
pillow>=9.0
img2pdf
